reorganize directory structure to better support multiple libraries inside a package
