+++ wombat_cmake/WombatInstallTargetsExtras.cmake
@@ -1,0 +1,4 @@
+# Copyright 2021 Azzollini Ilario, Gentilini Lorenzo, Soragna Alberto, Tazzari Roberto.
+# All Rights Reserved.
+
+include("${wombat_cmake_DIR}/WombatInstallTargets.cmake")
