+++ wombat_control/models/include/wombat_control/models/model_f_euler_dis.hpp
@@ -1,8 +1,8 @@
 // Copyright 2021 Azzollini Ilario, Gentilini Lorenzo, Soragna Alberto, Tazzari Roberto.
 // All Rights Reserved.
 
-#ifndef WOMBAT_CONTROL__MODEL__MODEL_F_EULER_DIS_HPP_
-#define WOMBAT_CONTROL__MODEL__MODEL_F_EULER_DIS_HPP_
+#ifndef WOMBAT_CONTROL__MODELS__MODEL_F_EULER_DIS_HPP_
+#define WOMBAT_CONTROL__MODELS__MODEL_F_EULER_DIS_HPP_
 
 #include "wombat_control/data_types/pose2d.hpp"
 
@@ -17,4 +17,4 @@ private:
   Pose2D m_current_pose;
 };
 
-#endif  // WOMBAT_CONTROL__MODEL__MODEL_F_EULER_DIS_HPP_
+#endif  // WOMBAT_CONTROL__MODELS__MODEL_F_EULER_DIS_HPP_
